requirements.txt
langchain-anthropic>=0.1.1
langchain-core>=0.2.22,<0.3.0
langgraph>=0.1.11
mysql-connector-python>=8.3.0
python-dotenv>=1.0.1
requests>=2.31.0
streamlit>=1.30.0